requirements.txt
fastapi
uvicorn
automata-lib
automata-lib[visual]
coloraide
pygraphviz
fastapi[standard]
pytest